hf_ocr_space/requirements.txt
pytesseract
numpy
Pillow
